+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.5.8-SNAPSHOT]
 
+- Pane builder (https://github.com/edvin/tornadofx/issues/208)
 - Iterable<Node>.style will apply styles to all elements in collection
 - Added `Node.alignment` property that knows how to apply alignment depending on the parent
 - Added `Node.margin` property that knows how to apply margin depending on the parent